Update source code documentation
This commit provides updated source code documentation and some enhancements.

UPDATED: License information.

ADDED: Source code comments. Try to build documentation with Xcode 13 DocC tool to have a bird's-eye view of the public API of the package.

ADDED: Support of system colors Mint, Cyan, and Brown.

CHANGED: Specification of system Teal color in according to the apple specification,

UPDATED: Unit tests.

UPDATED: README.

UPDATED: [PerseusDarkModeSingle.swift] file.

ENHANCED:  Instead of UIWindowAdaptable removed there is AppearanceService.processTraitCollectionDidChange(_:) is provided to match Dark Mode with System Appearance.

ENHANCED: DarkModeImageView can be configured via Interface Builder using Attributes Inspector.
