+++ Package.swift
@@ -1,8 +1,14 @@
 // swift-tools-version:5.3
-//
-// Package.swift
-//
-// Copyright © 2022 Mikhail Zhigulin. All rights reserved.
+/* Package.swift
+ 
+ Created by Mikhail Zhigulin in 7530.
+ 
+ Copyright © 7530 Mikhail Zhigulin of Novosibirsk.
+ All rights reserved.
+ 
+ Abstract:
+ Package manifest for Dark Mode.
+*/
 
 import PackageDescription
