--- Sources/AdaptedSystemUI/ColorHelper.swift
@@ -1,43 +1,0 @@
-//
-// ColorHelper.swift
-// AdaptedSystemUI
-//
-// Copyright © 2022 Mikhail Zhigulin. All rights reserved.
-
-#if !os(macOS)
-import UIKit
-#endif
-
-public func rgba255(_ red  : CGFloat,
-                    _ green: CGFloat,
-                    _ blue : CGFloat,
-                    _ alpha: CGFloat = 1.0) -> UIColor
-{
-    return UIColor(red: red/255, green: green/255, blue: blue/255, alpha: alpha)
-}
-
-public extension UIColor
-{
-    var RGBA255: (red: CGFloat, green: CGFloat, blue: CGFloat, alpha: CGFloat)
-    {
-        var red: CGFloat = 0
-        var green: CGFloat = 0
-        var blue: CGFloat = 0
-        var alpha: CGFloat = 0
-        
-        getRed(&red, green: &green, blue: &blue, alpha: &alpha)
-        
-        return (red*255, green*255, blue*255, alpha)
-    }
-}
-
-extension UIColor
-{
-    ///
-    /// Used only for unit testing to take control of selecting a color scheme.
-    ///
-    /// If it is true and if #available(iOS 13.0, *) is also true, the branch with the color scheme for iOS 13 and higher will be selected,
-    /// otherwise if it is false the color scheme for iOS 12 and lower will be selected at any way.
-    ///
-    internal static var _iOS13InUseAndHigherOnly: Bool = true
-}
