+++ Sources/AdaptedSystemUI/Implementation/SystemColorsAdapted.swift
@@ -1,8 +1,12 @@
 //
-// SystemColorsAdapted.swift
-// AdaptedSystemUI
+//  SystemColorsAdapted.swift
+//  AdaptedSystemUI
 //
-// Copyright © 2022 Mikhail Zhigulin. All rights reserved.
+//  Created by Mikhail Zhigulin in 7530.
+//
+//  Copyright © 7530 Mikhail Zhigulin of Novosibirsk.
+//  All rights reserved.
+//
 
 #if !os(macOS)
 import UIKit
@@ -10,51 +14,21 @@ import UIKit
 
 import PerseusDarkMode
 
-// MARK: System colors
-
-public protocol UISystemColorsAdapted
+/// Implementation of adaptation of system colors.
+///
+/// Each adaptation is either native SDK color or self-made.
+///
+/// For those iOS versions where a color is not released self-made adaptation is used.
+extension UIColor: SystemColorProtocol
 {
-    // MARK: - Default set of system colors
-    
-    ///
-    /// System colors.
-    ///
-    static var systemRed_Adapted   : UIColor { get }
-    static var systemOrange_Adapted: UIColor { get }
-    static var systemYellow_Adapted: UIColor { get }
-    static var systemGreen_Adapted : UIColor { get }
-    static var systemMint_Adapted  : UIColor { get }
-    static var systemTeal_Adapted  : UIColor { get }
-    static var systemCyan_Adapted  : UIColor { get }
-    static var systemBlue_Adapted  : UIColor { get }
-    static var systemIndigo_Adapted: UIColor { get }
-    static var systemPurple_Adapted: UIColor { get }
-    static var systemPink_Adapted  : UIColor { get }
-    static var systemBrown_Adapted : UIColor { get }
-    
-    ///
-    /// System gray group.
-    ///
-    static var systemGray_Adapted  : UIColor { get }
-    static var systemGray2_Adapted : UIColor { get }
-    static var systemGray3_Adapted : UIColor { get }
-    static var systemGray4_Adapted : UIColor { get }
-    static var systemGray5_Adapted : UIColor { get }
-    static var systemGray6_Adapted : UIColor { get }
-}
-
-extension UIColor: UISystemColorsAdapted
-{
-    ///
-    /// SYSTEM RED
-    ///
+    /// Red.
     public static var systemRed_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(255, 59, 48) : rgba255(255, 69, 58)
+            
+            rgba255(255, 59, 48) : rgba255(255, 69, 58)
             
             return color
         }
@@ -62,16 +36,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemRed
     }
     
-    ///
-    /// SYSTEM ORANGE
-    ///
+    /// Orange.
     public static var systemOrange_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(255, 149, 0) : rgba255(255, 159, 10)
+            
+            rgba255(255, 149, 0) : rgba255(255, 159, 10)
             
             return color
         }
@@ -79,16 +51,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemOrange
     }
     
-    ///
-    /// SYSTEM YELLOW
-    ///
+    /// Yeallow.
     public static var systemYellow_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(255, 204, 0) : rgba255(255, 214, 10)
+            
+            rgba255(255, 204, 0) : rgba255(255, 214, 10)
             
             return color
         }
@@ -96,16 +66,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemYellow
     }
     
-    ///
-    /// SYSTEM GREEN
-    ///
+    /// Green.
     public static var systemGreen_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(52, 199, 89) : rgba255(48, 209, 88)
+            
+            rgba255(52, 199, 89) : rgba255(48, 209, 88)
             
             return color
         }
@@ -113,46 +81,46 @@ extension UIColor: UISystemColorsAdapted
         return .systemGreen
     }
     
-    ///
-    /// SYSTEM MINT
-    ///
+    /// Mint.
     public static var systemMint_Adapted: UIColor
     {
-        //
-        // .systemMint is not available in SDK
-        //
-        let color = AppearanceService.shared.Style == .light ?
+        guard #available(iOS 15.0, *), _iOS13InUseAndHigherOnly else
+        {
+            let color = AppearanceService.shared.Style == .light ?
             
             rgba255(0, 199, 190) : rgba255(102, 212, 207)
+            
+            return color
+        }
         
-        return color
+        return .systemMint
     }
     
-    ///
-    /// SYSTEM TEAL
-    ///
+    /// Teal.
     public static var systemTeal_Adapted: UIColor
     {
-        // Gives unexpected color in iOS 13 and higher.
+        // .systemTeal gives unexpected color
+        // in iOS 13, but meets in 15.
         //
         // Tested on iOS 13.7.
         //
-        // Specification RGBA: 48, 176, 199 in Light.
-        // Specification RGBA: 64, 200, 224 in Dark.
+        // Specification:
+        // RGBA: 48, 176, 199 in Light
+        // RGBA: 64, 200, 224 in Dark
         //
-        // But it doesn't meet SDK .systemTeal certain color value in iOS 13 and higher.
+        // Simulator iPhone 8 (iOS 13.7):
+        // RGBA: 90, 200, 250 in Light
+        // RGBA: 100, 210, 255 in Dark
         //
         // LINK to .systemTeal color specification:
         // https://developer.apple.com/design/human-interface-guidelines/ios/visual-design/color/
         //
-        // So, SDK version is used for sure.
-        // SDK version has been extracted from iOS 13.7.
         
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(90, 200, 250) : rgba255(100, 210, 255)
+            
+            rgba255(48, 176, 199) : rgba255(64, 200, 224)
             
             return color
         }
@@ -161,31 +129,29 @@ extension UIColor: UISystemColorsAdapted
         
     }
     
-    ///
-    /// SYSTEM CYAN
-    ///
+    /// Cyan.
     public static var systemCyan_Adapted: UIColor
     {
-        //
-        // .systemCyan is not available in SDK
-        //
-        let color = AppearanceService.shared.Style == .light ?
+        guard #available(iOS 15.0, *), _iOS13InUseAndHigherOnly else
+        {
+            let color = AppearanceService.shared.Style == .light ?
             
             rgba255(50, 173, 230) : rgba255(100, 210, 255)
+            
+            return color
+        }
         
-        return color
+        return .systemCyan
     }
     
-    ///
-    /// SYSTEM BLUE
-    ///
+    /// Blue.
     public static var systemBlue_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(0, 122, 255) : rgba255(10, 132, 255)
+            
+            rgba255(0, 122, 255) : rgba255(10, 132, 255)
             
             return color
         }
@@ -193,16 +159,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemBlue
     }
     
-    ///
-    /// SYSTEM INDIGO
-    ///
+    /// Indigo.
     public static var systemIndigo_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(88, 86, 214) : rgba255(94, 92, 230)
+            
+            rgba255(88, 86, 214) : rgba255(94, 92, 230)
             
             return color
         }
@@ -210,16 +174,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemIndigo
     }
     
-    ///
-    /// SYSTEM PURPLE
-    ///
+    /// Purple.
     public static var systemPurple_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(175, 82, 222) : rgba255(191, 90, 242)
+            
+            rgba255(175, 82, 222) : rgba255(191, 90, 242)
             
             return color
         }
@@ -227,16 +189,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemPurple
     }
     
-    ///
-    /// SYSTEM PINK
-    ///
+    /// Pink.
     public static var systemPink_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(255, 45, 85) : rgba255(255, 55, 95)
+            
+            rgba255(255, 45, 85) : rgba255(255, 55, 95)
             
             return color
         }
@@ -244,31 +204,29 @@ extension UIColor: UISystemColorsAdapted
         return .systemPink
     }
     
-    ///
-    /// SYSTEM BROWN
-    ///
+    /// Brown.
     public static var systemBrown_Adapted: UIColor
     {
-        //
-        // .systemBrown is not available in SDK
-        //
-        let color = AppearanceService.shared.Style == .light ?
+        guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
+        {
+            let color = AppearanceService.shared.Style == .light ?
             
             rgba255(162, 132, 94) : rgba255(172, 142, 104)
+            
+            return color
+        }
         
-        return color
+        return .systemBrown
     }
     
-    ///
-    /// SYSTEM GRAY
-    ///
+    /// Gray.
     public static var systemGray_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(142, 142, 147) : rgba255(142, 142, 147)
+            
+            rgba255(142, 142, 147) : rgba255(142, 142, 147)
             
             return color
         }
@@ -276,16 +234,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemGray
     }
     
-    ///
-    /// SYSTEM GRAY 2
-    ///
+    /// System gray 2.
     public static var systemGray2_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(174, 174, 178) : rgba255(99, 99, 102)
+            
+            rgba255(174, 174, 178) : rgba255(99, 99, 102)
             
             return color
         }
@@ -293,16 +249,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemGray2
     }
     
-    ///
-    /// SYSTEM GRAY 3
-    ///
+    /// System gray 3.
     public static var systemGray3_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(199, 199, 204) : rgba255(72, 72, 74)
+            
+            rgba255(199, 199, 204) : rgba255(72, 72, 74)
             
             return color
         }
@@ -310,16 +264,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemGray3
     }
     
-    ///
-    /// SYSTEM GRAY 4
-    ///
+    /// System gray 4.
     public static var systemGray4_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(209, 209, 214) : rgba255(58, 58, 60)
+            
+            rgba255(209, 209, 214) : rgba255(58, 58, 60)
             
             return color
         }
@@ -327,16 +279,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemGray4
     }
     
-    ///
-    /// SYSTEM GRAY 5
-    ///
+    /// System gray 5.
     public static var systemGray5_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(229, 229, 234) : rgba255(44, 44, 46)
+            
+            rgba255(229, 229, 234) : rgba255(44, 44, 46)
             
             return color
         }
@@ -344,16 +294,14 @@ extension UIColor: UISystemColorsAdapted
         return .systemGray5
     }
     
-    ///
-    /// SYSTEM GRAY 6
-    ///
+    /// System gray 6.
     public static var systemGray6_Adapted: UIColor
     {
         guard #available(iOS 13.0, *), _iOS13InUseAndHigherOnly else
         {
             let color = AppearanceService.shared.Style == .light ?
-                
-                rgba255(242, 242, 247) : rgba255(28, 28, 30)
+            
+            rgba255(242, 242, 247) : rgba255(28, 28, 30)
             
             return color
         }
