+++ Sources/AdaptedSystemUI/Requirements/SemanticColorProtocol.swift
@@ -1,0 +1,143 @@
+//
+//  SemanticColorProtocol.swift
+//  AdaptedSystemUI
+//
+//  Created by Mikhail Zhigulin in 7530.
+//
+//  Copyright © 7530 Mikhail Zhigulin of Novosibirsk.
+//  All rights reserved.
+//
+
+#if !os(macOS)
+import UIKit
+#endif
+
+/// Represents semantic colors requirements.
+public protocol SemanticColorProtocol
+{
+    // MARK: - FOREGROUND CONTENT
+    
+    // MARK: - Label Colors
+    
+    /// Label.
+    ///
+    /// - Light: 0, 0, 0
+    /// - Dark: 255, 255, 255
+    static var label_Adapted               : UIColor { get }
+    
+    /// Secondary label.
+    ///
+    /// - Light: 60, 60, 67, 0.6
+    /// - Dark: 235, 235, 245, 0.6
+    static var secondaryLabel_Adapted      : UIColor { get }
+    
+    /// Tertiary label.
+    ///
+    /// - Light: 60, 60, 67, 0.3
+    /// - Dark: 235, 235, 245, 0.3
+    static var tertiaryLabel_Adapted       : UIColor { get }
+    
+    /// Quaternary label.
+    ///
+    /// - Light: 60, 60, 67, 0.18
+    /// - Dark: 235, 235, 245, 0.16
+    static var quaternaryLabel_Adapted     : UIColor { get }
+    
+    // MARK: - Text Colors
+    
+    /// Placeholder text.
+    ///
+    /// - Light: 60, 60, 67, 0.3
+    /// - Dark: 235, 235, 245, 0.3
+    static var placeholderText_Adapted     : UIColor { get }
+    
+    // MARK: - Separator Colors
+    
+    /// Separator.
+    ///
+    /// - Light: 60, 60, 67, 0.29
+    /// - Dark: 84, 84, 88, 0.6
+    static var separator_Adapted           : UIColor { get }
+    
+    /// Opaque separator.
+    ///
+    /// - Light: 198, 198, 200
+    /// - Dark: 56, 56, 58
+    static var opaqueSeparator_Adapted     : UIColor { get }
+    
+    // MARK: - Link Color
+    
+    /// Link.
+    ///
+    /// - Light: 0, 122, 255
+    /// - Dark: 9, 132, 255
+    static var link_Adapted                : UIColor { get }
+    
+    // MARK: - Fill Colors
+    
+    /// System fill.
+    ///
+    /// - Light: 120, 120, 128, 0.2
+    /// - Dark: 120, 120, 128, 0.36
+    static var systemFill_Adapted          : UIColor { get }
+    
+    /// Secondary system fill.
+    ///
+    /// - Light: 120, 120, 128, 0.16
+    /// - Dark: 120, 120, 128, 0.32
+    static var secondarySystemFill_Adapted : UIColor { get }
+    
+    /// Tertiary system fill.
+    ///
+    /// - Light: 118, 118, 128, 0.12
+    /// - Dark: 118, 118, 128, 0.24
+    static var tertiarySystemFill_Adapted  : UIColor { get }
+    
+    /// Quaternary system fill.
+    ///
+    /// - Light: 116, 116, 128, 0.08
+    /// - Dark: 118, 118, 128, 0.18
+    static var quaternarySystemFill_Adapted: UIColor { get }
+    
+    // MARK: - BACKGROUND CONTENT
+    
+    // MARK: - Standard
+    
+    /// System background.
+    ///
+    /// - Light: 255, 255, 255
+    /// - Dark: 28, 28, 30
+    static var systemBackground_Adapted                : UIColor { get }
+    
+    /// Secondary system background.
+    ///
+    /// - Light: 242, 242, 247
+    /// - Dark: 44, 44, 46
+    static var secondarySystemBackground_Adapted       : UIColor { get }
+    
+    /// Tertiary system background.
+    ///
+    /// - Light: 255, 255, 255
+    /// - Dark: 58, 58, 60
+    static var tertiarySystemBackground_Adapted        : UIColor { get }
+    
+    // MARK: - Grouped
+    
+    /// System grouped background.
+    ///
+    /// - Light: 242, 242, 247
+    /// - Dark: 28, 28, 30
+    static var systemGroupedBackground_Adapted         : UIColor { get }
+    
+    /// Secondary system grouped background.
+    ///
+    /// - Light: 255, 255, 255
+    /// - Dark: 44, 44, 46
+    static var secondarySystemGroupedBackground_Adapted: UIColor { get }
+    
+    /// Tertiary system grouped background.
+    ///
+    /// - Light: 242, 242, 247
+    /// - Dark: 58, 58, 60
+    static var tertiarySystemGroupedBackground_Adapted : UIColor { get }
+}
