+++ Sources/PerseusDarkMode/DarkMode/Constants.swift
@@ -1,11 +1,23 @@
 //
-// Constants.swift
-// PerseusDarkMode
+//  Constants.swift
+//  PerseusDarkMode
 //
-// Copyright © 2022 Mikhail Zhigulin. All rights reserved.
+//  Created by Mikhail Zhigulin in 7530.
+//
+//  Copyright © 7530 Mikhail Zhigulin of Novosibirsk.
+//  All rights reserved.
+//
 
 import Foundation
 
+/// User defaults key for AppearanceService.DarkModeUserChoice variable.
 public let DARK_MODE_USER_CHOICE_OPTION_KEY = "DarkModeUserChoiceOptionKey"
+
+/// Default value for AppearanceService.DarkModeUserChoice variable.
 public let DARK_MODE_USER_CHOICE_DEFAULT = DarkModeOption.auto
+
+/// Default valur for AppearanceService.shared.Style.
 public let DARK_MODE_STYLE_DEFAULT = AppearanceStyle.light
+
+/// Name of DarkMode.StyleObservable variable used in KVO
+public let OBSERVERED_VARIABLE_NAME = "StyleObservable"
