+++ Sources/PerseusDarkMode/DarkMode/DarkModeObserver.swift
@@ -1,36 +1,68 @@
 //
-// DarkModeObserver.swift
-// PerseusDarkMode
+//  DarkModeObserver.swift
+//  PerseusDarkMode
 //
-// Copyright © 2022 Mikhail Zhigulin. All rights reserved.
+//  Created by Mikhail Zhigulin in 7530.
+//
+//  Copyright © 7530 Mikhail Zhigulin of Novosibirsk.
+//  All rights reserved.
+//
 
 #if !os(macOS)
 import UIKit
 #endif
 
+/// Represents an object responsible for making run the code if the app's appearance style changed.
+///
+/// It connects to shared Dark Mode instance using appearance service.
+/// Also, it takes action every time when the app's appearance style changed.
+///
+/// The action can be specified with initialization as the passed closue and after initialization by assigning the action property as well.
 public class DarkModeObserver: NSObject
 {
-    public var action         : ((_ newStyle: AppearanceStyle)->Void)?
-    public let objectToObserve: DarkMode
+    /// Closure to perform if the app's appearance style changed.
+    public var action: ((_ newStyle: AppearanceStyle)->Void)?
     
-    public init(_ value: DarkMode)
+    /// The reference of the object to be obsevered.
+    private(set) var objectToObserve = AppearanceService.shared
+    
+    /// Initializer by default.
+    ///
+    /// It creates the instance with no action specified if the app's appearance style changed.
+    ///
+    /// Then, give it a closure to run the code if the app's appearance style changed.
+    public override init()
     {
-        objectToObserve = value
         super.init()
         
         objectToObserve.addObserver(self,
-                                    forKeyPath: "StyleObservable",
+                                    forKeyPath: OBSERVERED_VARIABLE_NAME,
                                     options   : .new,
                                     context   : nil)
     }
     
+    /// Initializer with parameters.
+    ///
+    /// Pass a closure to specify the action to be taken if the app's appearance style changed.
+    public init(_ action: @escaping ((_ newStyle: AppearanceStyle)->Void))
+    {
+        super.init()
+        
+        self.action = action
+        objectToObserve.addObserver(self,
+                                    forKeyPath: OBSERVERED_VARIABLE_NAME,
+                                    options   : .new,
+                                    context   : nil)
+    }
+    
+    /// Called each time when the abserved varible changed.
     public override func observeValue(forKeyPath keyPath: String?,
                                       of object         : Any?,
                                       change            : [NSKeyValueChangeKey : Any]?,
                                       context           : UnsafeMutableRawPointer?)
     {
         guard
-            keyPath == "StyleObservable",
+            keyPath == OBSERVERED_VARIABLE_NAME,
             let style = change?[.newKey],
             let styleRawValue = style as? Int,
             let newStyle = AppearanceStyle.init(rawValue: styleRawValue)
@@ -41,6 +73,6 @@ public class DarkModeObserver: NSObject
     
     deinit
     {
-        objectToObserve.removeObserver(self, forKeyPath: "StyleObservable")
+        objectToObserve.removeObserver(self, forKeyPath: OBSERVERED_VARIABLE_NAME)
     }
 }
