+++ Sources/PerseusDarkMode/DarkMode/DarkModeOption.swift
@@ -1,17 +1,27 @@
 //
-// DarkModeOption.swift
-// PerseusDarkMode
+//  DarkModeOption.swift
+//  PerseusDarkMode
 //
-// Copyright © 2022 Mikhail Zhigulin. All rights reserved.
+//  Created by Mikhail Zhigulin in 7530.
+//
+//  Copyright © 7530 Mikhail Zhigulin of Novosibirsk.
+//  All rights reserved.
+//
 
 import Foundation
 
+/// Represents a list of possible values for Dark Mode option.
+///
+/// - AUTO means the choice is made by System.
+/// - ON means the app's appearance style should be Dark at any case.
+/// - OFF means the app's appearance style should be Light at any case.
 public enum DarkModeOption: Int, CustomStringConvertible
 {
     case auto = 0
     case on   = 1
     case off  = 2
     
+    /// Textual representation of the current value of the option.
     public var description: String
     {
         switch self
