--- Sources/PerseusDarkMode/SetupByDefault.swift
@@ -1,28 +1,0 @@
-//
-// SetupByDefault.swift
-// PerseusDarkMode
-//
-// Copyright © 2022 Mikhail Zhigulin. All rights reserved.
-
-#if !os(macOS)
-import UIKit
-#endif
-
-public extension UIResponder { var DarkMode: DarkModeProtocol { AppearanceService.shared } }
-
-public class UIWindowAdaptable: UIWindow
-{
-    public override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?)
-    {
-        if AppearanceService._changeManually { return }
-        
-        super.traitCollectionDidChange(previousTraitCollection)
-        
-        guard #available(iOS 13.0, *),
-              let previousSystemStyle = previousTraitCollection?.userInterfaceStyle,
-              previousSystemStyle.rawValue != DarkMode.SystemStyle.rawValue
-        else { return }
-        
-        AppearanceService._systemCalledMakeUp()
-    }
-}
